pyproject.toml
[tool.poetry]
name = "myuplink"
version = "0.1.0"
description = "MyUplink is a simple library to interact with the internal Uplink APIs that are better than that the official ones"
authors = ["Stefano Mosconi <stefano.mosconi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
pandas = "^2.2.1"
matplotlib = "^3.8.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
